+++ lib/features/down_for_maintenance/data/services/mock_check_server_status_service.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter_template/features/down_by_maintenance/domain/entities/check_result.dart';
-import 'package:flutter_template/features/down_by_maintenance/domain/services/check_server_status_service.dart';
+import 'package:flutter_template/features/down_for_maintenance/domain/entities/check_result.dart';
+import 'package:flutter_template/features/down_for_maintenance/domain/services/check_server_status_service.dart';
 
 /// Implementation of [ICheckServerStatusService] that mocks server status.
 class MockCheckServerStatusService implements ICheckServerStatusService {
