+++ lib/features/down_for_maintenance/data/services/rest_check_server_status_service.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter_template/features/down_by_maintenance/domain/entities/check_result.dart';
-import 'package:flutter_template/features/down_by_maintenance/domain/services/check_server_status_service.dart';
+import 'package:flutter_template/features/down_for_maintenance/domain/entities/check_result.dart';
+import 'package:flutter_template/features/down_for_maintenance/domain/services/check_server_status_service.dart';
 
 /// Implementation of [ICheckServerStatusService] that uses REST API to check server status.
 class RestCheckServerStatusService implements ICheckServerStatusService {
