+++ lib/features/down_for_maintenance/domain/entities/check_result.dart
