+++ lib/features/down_for_maintenance/domain/services/check_server_status_service.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter_template/features/down_by_maintenance/domain/entities/check_result.dart';
+import 'package:flutter_template/features/down_for_maintenance/domain/entities/check_result.dart';
 
 /// Interface for checking server status.
 abstract interface class ICheckServerStatusService {
